Dagger: Scopes for Fragments
