+++ app/src/main/java/com/godamy/myplayer/di/AppComponent.kt
@@ -1,8 +1,10 @@
 package com.godamy.myplayer.di
 
 import android.app.Application
-import com.godamy.myplayer.ui.detail.DetailViewModelFactory
-import com.godamy.myplayer.ui.main.MainViewModelFactory
+import com.godamy.myplayer.ui.detail.di.DetailFragmentComponent
+import com.godamy.myplayer.ui.detail.di.DetailFragmentModule
+import com.godamy.myplayer.ui.main.di.MainFragmentComponent
+import com.godamy.myplayer.ui.main.di.MainFragmentModule
 import dagger.BindsInstance
 import dagger.Component
 import javax.inject.Singleton
@@ -12,14 +14,13 @@ import javax.inject.Singleton
     modules = [
         AppModule::class,
         DataModule::class,
-        UserCaseModule::class,
-        ViewModelModule::class
+        UserCaseModule::class
     ]
 )
 interface AppComponent {
 
-    val mainViewModelFactory: MainViewModelFactory
-    val detailViewModelFactory: DetailViewModelFactory
+    fun plus(module: MainFragmentModule): MainFragmentComponent
+    fun plus(module: DetailFragmentModule): DetailFragmentComponent
 
     @Component.Factory
     interface Factory {
