+++ app/src/main/java/com/godamy/myplayer/ui/detail/di/DetailFragmentComponent.kt
@@ -1,0 +1,9 @@
+package com.godamy.myplayer.ui.detail.di
+
+import com.godamy.myplayer.ui.detail.DetailViewModelFactory
+import dagger.Subcomponent
+
+@Subcomponent(modules = [DetailFragmentModule::class])
+interface DetailFragmentComponent {
+    val detailViewModelFactory: DetailViewModelFactory
+}
