+++ app/src/main/java/com/godamy/myplayer/ui/main/di/MainFragmentComponent.kt
@@ -1,0 +1,9 @@
+package com.godamy.myplayer.ui.main.di
+
+import com.godamy.myplayer.ui.main.MainViewModelFactory
+import dagger.Subcomponent
+
+@Subcomponent(modules = [MainFragmentModule::class])
+interface MainFragmentComponent {
+    val mainViewModelFactory: MainViewModelFactory
+}
