+++ app/src/main/java/com/godamy/myplayer/ui/main/di/MainFragmentModule.kt
@@ -1,0 +1,17 @@
+package com.godamy.myplayer.ui.main.di
+
+import com.godamy.myplayer.ui.main.MainViewModelFactory
+import com.godamy.myplayer.usecases.GetPopularMoviesUserCase
+import com.godamy.myplayer.usecases.RequestPopularMoviesUseCase
+import dagger.Module
+import dagger.Provides
+
+@Module
+object MainFragmentModule {
+
+    @Provides
+    fun provideMainViewModelFactory(
+        getPopularMoviesUserCase: GetPopularMoviesUserCase,
+        requestPopularMoviesUseCase: RequestPopularMoviesUseCase
+    ) = MainViewModelFactory(getPopularMoviesUserCase, requestPopularMoviesUseCase)
+}
